% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello48-package.R
\name{hello48}
\alias{hello48}
\title{It does some stuff}
\description{
It does some stuff
}
\section{Support}{


The development repository for the **hello48** package is found at
\url{https://github.com/noamross/hello48}.  Please file
bug reports or other feedback at \url{https://github.com/noamross/hello48/issues}
}

\author{
Noam Ross \email{noam.ross@gmail.com}
}
\keyword{package}
